fix: address PRRT_kwDOSJAM6s6RPzO4 — split oversized hosted validation test part
